requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
easyocr==1.7.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
gunicorn==20.1.0
python-multipart==0.0.7